Use existing location
