+++ spec/system/report_interface_spec.rb
@@ -28,4 +28,19 @@ RSpec.describe 'report interface', :attachment_processing, :js, :streaming do
     page.scroll_to(page.find('.batch-table__row'))
     expect(page).to have_css('.spoiler-button__overlay__label')
   end
+
+  it 'marks a report resolved from the show page actions area' do
+    visit admin_report_path(report)
+
+    # puts page.body
+
+    expect { resolve_report }
+      .to change { report.reload.action_taken_at }.to(be_present).from(nil)
+  end
+
+  def resolve_report
+    within '.report-actions' do
+      click_on I18n.t('admin.reports.mark_as_resolved')
+    end
+  end
 end
